python_Version/requirements.txt
selenium==4.8.0
pyyaml==6.0